requirements.txt
discord.py
uvloop
python-dotenv
regex
tortoise-orm
